moved update trade route test
